4.21

+++ controllers/blogs.js
@@ -4,13 +4,7 @@ const User = require('../models/user')
 
 
 const jwt = require('jsonwebtoken')
-const getTokenFrom2 = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
+
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -68,11 +62,25 @@ blogsRouter.post('/',  async (request, response, next) => {
 
   
 blogsRouter.delete('/:id',  async (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then(() => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  console.log(user.id)
+  console.log(blog.user)
+  if ( blog.user.toString() === user.id.toString() ) {
+    Blog.findByIdAndRemove(request.params.id)
+      .then(() => {
+        response.status(204).end()
+      })
+      .catch(error => next(error))
+
+  }
+
 })
   
 blogsRouter.put('/:id', (request, response, next) => {
